chore: support hot-patch
Prior to this commit, the loop function will keep the process
locked in the beam version in use.
When hot-patched, the old beam will eventually be purged
causing the processes still running on the old version beam
to be killed.

This commit tries to use remote (M:F/A) calls to do the main
loop, so the process has a better chance to enter the new
version beam.

For idling processes though, if the keepalive interval is
infinity or if it's long enough, there is still a good chance
for it to get killed because there is no trigger for the
process to enter the new module.
  -- The idea is to let this case go, because restarting
an idling process is not a big deal, no on-going requests
to be affected anyway.